+++ src/main/java/com/example/electro/controller/mvc/CustomErrorController.java
@@ -1,11 +1,10 @@
-package com.example.electro.controller;
+package com.example.electro.controller.mvc;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.boot.web.servlet.error.ErrorController;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-
 
 @Controller
 public class CustomErrorController implements ErrorController {
