+++ src/main/java/com/example/electro/controller/mvc/OrderCustomerController.java
@@ -1,4 +1,4 @@
-package com.example.electro.controller;
+package com.example.electro.controller.mvc;
 
 import com.example.electro.customDetails.CustomUserDetails;
 import com.example.electro.dto.ErrorResponse;
