+++ src/main/java/com/example/electro/controller/mvc/UpdateProductController.java
@@ -1,6 +1,5 @@
-package com.example.electro.controller;
+package com.example.electro.controller.mvc;
 
-import com.example.electro.service.CategoryService;
 import com.example.electro.service.FilterService;
 import com.example.electro.service.ProductService;
 import org.springframework.stereotype.Controller;
