+++ src/main/java/com/example/electro/repository/OrderItemRepository.java
@@ -1,0 +1,6 @@
+package com.example.electro.repository;
+
+import com.example.electro.model.OrderItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {}
